Now connects successfully to the DB using the parameters entered in the initialConfigurationPage's form.
Also changed task_done to progress_update which passes the new progress
value as an argument. 

Also propogated the 'config_error' event back up to main and finally to
the client via socket.io to display the error to the user.

And did some work on the css for the form to make it look nicer.

TODO: should probably run through what we have so far, document it in
some form of javadac esque way , maybe rename some things to better
reflect what they are, and form those conventions and habits before
moving too much farther.