So I played around with sockerIO some more and extended the SocketIOCOnnectionTracker. My idea is that this class will be the utility through which the entire system actually sends emits and registers on callbacks. You may register for all clients or individually by id.
Internally the connection tracker associates a client_id with the
client, and creates a two way mapping between the client_id to the
real_id of the socket. THe client script remembers its id. Then whenever
the client receives the 'connect' callback it will send its remembered
client_id to the server. THe server then creates/updates the necessary
mapping between the real socket's id and the arbitrary client id. 

THe user can change this remembered id by editing the script in the
browser. If they change it to the id of an existing connection, they are
disconnected. If they decided to change it to a free id they will
receive a new connection and their application will likely not function.

This seems like a pretty solid way to go about things. And the id we
choose to tell the client about is arbitrary but using the first
socket.io id doesn't seem like it will be a problem unless we have
millions of concurrent users.