Fixed a persistent logic error with the client and the id.  Sending "undefined" as id no longer works well.  Changed default id to "defaultIdValue".
Found a new problem.  Can not upload two bots of the same name without
an error becuase the files are uploaded to the same temp file then moved
and multer does them async I guess.  Not sure about solution as of
writing this.  Likely to move when the files are moved.  Would like to
have them upload and moved based on input name field, but tried that
early.  Perhaps ill revisit that.