Lots of small tweaks. Created methods for hasInstanceExpired and getMillisendsUntilExpiration. Replaced all existence checks for testArenaInstances[id] with hasInstanceExpired(id) since it is more robust and wont try to do things on expired instances.
The milliseconds until expiration are now sent along with each request
for a gameState, move, or start and kill game. These are displayed in a
countdown fashion in the status message bar. Extended the status message
display queue logic so that it always displays each message for a
minimum of 500 ms if there are others waiting and a maximum of 3 seconds
if nothing else is on the queue. Will only slide up once nothing is on
the queue.