Added a "nextTurn" property to each game state. This will indicate which player is up next will will be helpful to determine if the human player is up and needs to enter input. Updated SaveTheIslandJSONFramework to indicate these changes. And finished the switchover from GameManager to ArenaGameManager.
Also changed the noGameRunning event I added in the last commit to
noStatesRemaining. This is now run only when the gameManager has exited
and there are no game states left in the server's queue. On this event
the gameStateRequester is stopped since it knows for sure there arent
going to be anymore game states. 

In the event that the GameManager crashes unexpectedly and never sends a
final game state, the client will hang in the running state, waiting for
the user to click kill game. They will however be told their are no more
game states coming.

I'm not sure how to handle this though because normal behavior has the
GameManager exit after sending the final game state which is potentially
long before the client actually processes that final game state. THe
nieve approach taken in the last commit caused the game animation to
halt once it got the noGameRunning event. I suppose an easy solution may
be to have the GameManager hold waiting to be explicitly killed by the
node server, but that's not ideal.