Implemented a prototype of the use of json-simple (https://code.google.com/p/json-simple/) for both parsing and json generation in ArenaGameManager. The JSON generated by it is received correctly and generates the animations just as the hardcoded JSON strings did. Since its pretty simple it may not be too bad to just generate the strings but as Steven and I found out it was quite a pain when you're missing one small thing and can't figure out what it is. JSON is actually quite string.
BotBattleApp is also now sending all the information needed in order to
run the Test arena. Which amounts to the paths to the compiled bots, and
the language (c++ or java) for convienience.

All that's left is to modify the GameManager code to output JSON
gameStates as the game runs.