Fixed two small issues with the JSON in getJSONStringForThisTurn. There was an extra curly brace after gameData and the semicolon at the end was not considered valid JSON. We'll need to get you set up with a working environment for node so you can test changes to the game logic moving forward. In the meantime you can use http://jsonlint.com/ to validate the JSON your generating or use that library but it seems like you already have most of it done.
You'll need to add the type and nextTurn attributes to this function as
well, which will require logic to determine whether its the initial
gameState, a midGame state, or the final gameState. The nextTurn looks
like it will be as simple as ((player+1)%2))+1. 

ALso in the event of the bot exiting on its own or a null or invalid
move sent. I suppose these should just be sent as gamestates with
"type":"final", and the debugging info placed in the debugInfo array
instead of in the animatableEvents as done now.

If you'd like to set up node yourself all you should need is to install
node js from the website.

You can use my hosted database by changing the config file at
"nodejs_applications/savedConfiguration.txt" to reference ambarket.info
as the hostname.

Then run using "node nodejs_applications/main.js" on command line. If
all went well it should now be accessible at https://localhost:6058

To test changes to your code all you need to do is copy the compiled
class files to
"nodejs_applications/local_storage/game_modules/SuperAwesomeFunGame/" 

Once the class files are there they will be run by BotBattleApp.