Added field for game javascript file and game resources to initial configuration page. Resources will be a zip file with any resources that are needed for the javascript file. These will ultimately be unzipped and made accessible at /game/[Game Name]/resources/.
Removed restriction that files uploaded to initial ocnfig have unique
names by just changing all references to .name to .originalname. 

Moved code out of onFileUploadStart that was checking for source to be a
java file and rules to be a pdf file. Its much cleaner just to upload
them regardless and check they match regex along with the rest of input.