Implemented the uploading and processing of game javascript and resources files in initial config. This required a fair amount of refactoring. I plan on doing some more testing before merging this branch in.
I also changed the GameModule object in the database to store the path
to the game manager source and class file directories. Currently the
class files are still placed in the source directory but I intend to
separate them. Just need to create the output directory first. 

Next think I'll do this week will probably be refactor the
BotBattleCompiler a bit, most likely creating a specialized method just
for compiling the game manager source directory.