Finished up most of what I started yesterday.
The rules file for any game in the database will now be served at
/game/[Game Name]/rules

The javascript file for that game will similarly be available at
/game/[Game Name]/javascript

ANy additional resources needed by the javascript file can be uploaded
in the form of a zip file to the initial config page. All the contents
of which, including subdirectories can be accessed at 
/game/[Game Name]/resources/[path relative to root of zip file]

The javascript file url is now sent through the locals object to
res.render instead of hardcoded. I plan on adding a link to the rules
file on the test arena page shortly.

ALso fixed logger so it will simply pass all arguments along to
console.log after processing the category and will print objects
correctly.