I made a lot of the changes I described in that issue so I decided to push to a new branch. I'm done for the night. The next step I think is to write a bot that doesn't almost always make an invalid move. so we can actually see a game go to completion.
Also we need to implement the client side human input. I was thinking we
should just make move, retreat, and attack buttons instead of left and
right so that it exactly matches what their bots will have to do. Then
just parse their selection into the structure of a move string and send
whatever they select to the GameManager for validation.

Also I think the tiles are actually being replaced I was just missing it
because it happens quick and are often replaced with the same time.
Would be nice to try to animate this a little but that is very low
priority.