Rearranged the game loop logic. I believe this works for all cases now. The main idea was to always either send a midGameState or an invalidMove JSON message in the loop. Only when you break out of the loop because game.isGameOver() is true do you send the finalGameState.
I removed the setOver and isGameWon from the game interface. The game
itself is responsible for determining when the game is over either in
validateMove() or in updateBoard(). And at that point of knowing the
game is over it sets a private variable indicating either which player
won, or which player was disqualified. This gameOverMessage is used in
the isGameOver() method, if its not null then the game has ended.

A made a bunch of other small tweaks and moved methods around as I was
trying to make sure all interface methods has @Override and make sure
the interface made sense and started to add javadoc style comments to
the interface.