Not really a change just wanted to comment.
So I completely missed that you already did the sending the player the
player number. I did it by just prepending it to their boards. Your way
of only sending once is probably better and is more like what steven
wanted too. We'll chose one and undo the other tomorrow.