Lots of small tweaks. The shuffle limit is now set to 20. I think this is high enough that any reasonable length game you should not have to worry about it, as long as your strategy isn't to just shuffle every time.
Also added the option to select either the preloaded bot or upload your
own. Currently this will only support a single hardcoded bot defined in
BotBattlePaths. Wouldn't be hard to support more and give them a
selectlist of options but its probably not worth it. I believe I've
testd all possible combinations of preloaded, custom, human games and
all seems to work well.

Also Added a stream of complete game boards to the test arena. Thought
this might be nice and was very quick to do.

Updated the rules to reflect all of this.