Fixed bug in initial config where if there was a significant delay between server and client the server would load BotBattleApp before the client ever figured out the configuartion was successful. Also this uncovered a copy and paste bug in BotBattleDatabase error handling code.
It seems that the most likely reason downloads do not work on phones is 

https://code.google.com/p/android/issues/detail?id=3492
http://stackoverflow.com/questions/23732736/bug-with-downloading-files-from-node-js-server-on-google-chrome-browser-on-andro

apparently it's just an issue with the fact that its HTTPS, and the
android download manager doesn't trust our self signed certificate and
won't even ask the user if they want to continue. It just hangs.