Finished up what I was doing to kill bots when game manager exits. It seems to work well.
Added human timeout of 5 minutes. If people complain we'll increase it.
Still need to test that this works.

Added getMoveTImout methods to the GameInterface they were static before
which was bad because they were used in player. Had to add a timeout
parameter to Player.getMove for this to work.